+++ db/migrate/20190611094629_create_presenting_complaints.rb
@@ -2,8 +2,8 @@ class CreatePresentingComplaints < ActiveRecord::Migration[5.2]
   def change
     create_table    :presenting_complaints, :primary_key => :presenting_complaint_id do |t|
       t.bigint          :encounter_id
-      t.integer        :concept_id
-      t.integer        :value_coded
+      t.bigint        :concept_id
+      t.bigint        :value_coded
       t.boolean       :voided, null: false, default: 0
       t.bigint          :voided_by
       t.datetime     :voided_date
